requirement.txt
aiosignal==1.2.0
async-timeout==4.0.2
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.15.30
awscrt==0.19.19
Babel==2.9.1
boto3==1.35.72
botocore==1.35.76
certifi==2022.6.15
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
chevron==0.13.1
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
dbus-python==1.2.18
distro==1.5.0
docutils==0.16
docx==0.2.4
ec2-hibinit-agent==1.0.8
et_xmlfile==2.0.0
frozenlist==1.3.0
gitdb==4.0.11
GitPython==3.1.41
gpg==1.15.1
graphql-core==3.2.1
idna==3.3
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
libcomps==0.1.20
lockfile==0.12.2
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.0.2
netifaces==0.10.6
numpy==2.0.2
oauthlib==3.0.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
pycparser==2.20
PyPDF2==3.0.1
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-docx==1.1.2
python-pptx==1.0.2
pytz==2022.7.1
PyYAML==6.0.1
release-notification==1.2
requests==2.25.1
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
s3transfer==0.10.4
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.16.0
smmap==5.0.1
support-info==1.0
systemd-python==235
trufflehog3==3.0.10
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.9
utils==1.0.2
wcwidth==0.2.5
XlsxWriter==3.2.0
yarl==1.7.2